+++ Contract.java
@@ -1,5 +1,7 @@
 package taxcalculator;
 
 public interface Contract {
+    String getName();
+
     TaxBreakdown calculate(double grossIncome);
 }